requirements.txt
nameparser
sqlalchemy
pymysql
requests
selenium
python-dotenv
cryptography
demoji
beautifulsoup4
lastpass-python
fuzzywuzzy
python-Levenshtein
google-cloud-secret-manager
google-cloud-pubsub
gspread
pandas